adopter_form_completed_and_display_lists_of_adopted_pets

+++ db/migrate/20231102111940_add_pet_id_to_pet_adoptions.rb
@@ -1,5 +1,5 @@
 class AddPetIdToPetAdoptions < ActiveRecord::Migration[7.1]
   def change
-    add_column :pet_adoptions, :pet_id, :integer
+    add_column :pet_adoptions, :pet_id, :bigint
   end
 end
